+++ apps/client/src/view/templates/TemplateListView.vue
@@ -1,12 +1,23 @@
 <script lang="ts" setup>
+import {
+  type DigitalProductDocumentStatusDtoType,
+  type PagingParamsDto,
+  type DigitalProductDocumentDto,
+  DigitalProductDocumentStatusDto,
+} from "@open-dpp/dto";
 import { onMounted, ref } from "vue";
 import { useI18n } from "vue-i18n";
 import { useRoute, useRouter } from "vue-router";
-import DppTable from "../../components/DppTable.vue";
+import DigitalProductDocumentTable from "../../components/digital-product-document/DigitalProductDocumentTable.vue";
 import TemplateCreateDialog from "../../components/template/TemplateCreateDialog.vue";
 import { useExportImport } from "../../composables/export-import.ts";
 import { useTemplates } from "../../composables/templates.ts";
 import apiClient from "../../lib/api-client.ts";
+import { usePagination } from "../../composables/pagination.ts";
+import { useDigitalProductDocumentFilter } from "../../composables/digital-product-document-filter.ts";
+import DigitalProductDocumentStatusChangeMenu from "../../components/digital-product-document/DigitalProductDocumentStatusChangeMenu.vue";
+import { useDigitalProductDocument } from "../../composables/digital-product-document.ts";
+import { DigitalProductDocumentType } from "../../lib/digital-product-document.ts";
 
 const route = useRoute();
 const router = useRouter();
@@ -20,21 +31,32 @@ function changeQueryParams(newQuery: Record<string, string | undefined>) {
   });
 }
 
+const { createTemplate, templates, loading, fetchTemplates } = useTemplates();
+
+const { deleteDPD, publish, restore, archive } = useDigitalProductDocument(
+  DigitalProductDocumentType.Template,
+);
+
+const { status, changeStatus } = useDigitalProductDocumentFilter();
+
+function fetchCallback(pagingParams: PagingParamsDto) {
+  return fetchTemplates(pagingParams, status.value ? { status: [status.value] } : undefined);
+}
 const {
-  createTemplate,
   resetCursor,
   hasPrevious,
   hasNext,
   previousPage,
   nextPage,
   currentPage,
-  templates,
-  loading,
-  init,
-} = useTemplates({
-  changeQueryParams,
+  reloadCurrentPage,
+} = usePagination({
   initialCursor: route.query.cursor ? String(route.query.cursor) : undefined,
+  limit: 10,
+  fetchCallback,
+  changeQueryParams,
 });
+
 const { t } = useI18n();
 
 const createDialogVisible = ref(false);
@@ -57,24 +79,50 @@ const {
   importErrorKey: "common.templateImportFailed",
 });
 
+async function onDeleteButtonClicked(item: DigitalProductDocumentDto) {
+  await deleteDPD(item.id, reloadCurrentPage);
+}
+
+async function onPublishButtonClicked(item: DigitalProductDocumentDto) {
+  await publish(item.id);
+  await reloadCurrentPage();
+}
+
+async function onArchiveButtonClicked(item: DigitalProductDocumentDto) {
+  await archive(item.id);
+  await reloadCurrentPage();
+}
+
+async function onRestoreButtonClicked(item: DigitalProductDocumentDto) {
+  await restore(item.id);
+  await reloadCurrentPage();
+}
+
+async function onSelectedStatusChange(newStatus: DigitalProductDocumentStatusDtoType | undefined) {
+  await changeStatus(newStatus);
+  await resetCursor();
+}
+
 onMounted(async () => {
-  await init();
+  await nextPage();
 });
 </script>
 
 <template>
-  <DppTable
+  <DigitalProductDocumentTable
     key="templates-list"
     :has-previous="hasPrevious"
     :has-next="hasNext"
     :current-page="currentPage"
     :items="templates ? templates.result : []"
     :loading="loading"
-    :title="t('templates.label')"
+    :title="t('templates.label', 2)"
     @reset-cursor="resetCursor"
     @create="createDialogVisible = true"
     @next-page="nextPage"
     @previous-page="previousPage"
+    :selected-status="status"
+    @update:selected-status="onSelectedStatusChange"
   >
     <template #headerActions>
       <Button :label="t('common.add')" @click="createDialogVisible = true" />
@@ -88,26 +136,43 @@ onMounted(async () => {
         @select="onTemplateFileSelect"
       />
     </template>
-    <template #actions="{ passport, editItem }">
+    <template #actions="{ item, goToItem }">
       <Button
+        v-if="status !== DigitalProductDocumentStatusDto.Archived"
         icon="pi pi-pencil"
         severity="primary"
         :aria-label="t('common.edit')"
         :title="t('common.edit')"
-        @click="editItem(passport)"
+        @click="goToItem(item)"
+      />
+      <Button
+        v-if="status === DigitalProductDocumentStatusDto.Archived"
+        icon="pi pi-eye"
+        severity="primary"
+        :aria-label="t('common.view')"
+        :title="t('common.view')"
+        @click="goToItem(item)"
       />
       <Button
         icon="pi pi-download"
         severity="secondary"
         :aria-label="t('common.exportTemplate')"
         :title="t('common.exportTemplate')"
-        @click="exportTemplate(passport.id)"
+        @click="exportTemplate(item.id)"
+      />
+      <DigitalProductDocumentStatusChangeMenu
+        :item="item"
+        @on-delete-clicked="onDeleteButtonClicked"
+        @on-publish-clicked="onPublishButtonClicked"
+        @on-archive-clicked="onArchiveButtonClicked"
+        @on-restore-clicked="onRestoreButtonClicked"
       />
     </template>
-  </DppTable>
+  </DigitalProductDocumentTable>
   <TemplateCreateDialog
     v-if="createDialogVisible"
     v-model="createDialogVisible"
     :create-template="createTemplate"
   />
+  <ConfirmDialog />
 </template>