+++ apps/main/src/templates/infrastructure/template.repository.spec.ts
@@ -1,21 +1,27 @@
 import type { TestingModule } from "@nestjs/testing";
+import { Test } from "@nestjs/testing";
 import { randomUUID } from "node:crypto";
 import { expect } from "@jest/globals";
-import { MongooseModule } from "@nestjs/mongoose";
-import { Test } from "@nestjs/testing";
+import { getModelToken, MongooseModule } from "@nestjs/mongoose";
 
 import { EnvModule, EnvService } from "@open-dpp/env";
+import { Model, Model as MongooseModel } from "mongoose";
 import { AasModule } from "../../aas/aas.module";
 import { Environment } from "../../aas/domain/environment";
 import { generateMongoConfig } from "../../database/config";
+import {
+  DigitalProductDocumentStatus,
+  DigitalProductDocumentStatusChange,
+} from "../../digital-product-document/domain/digital-product-document-status";
 import { encodeCursor, Pagination } from "../../pagination/pagination";
 import { PagingResult } from "../../pagination/paging-result";
 import { Template } from "../domain/template";
 import { TemplateRepository } from "./template.repository";
-import { TemplateDoc, TemplateSchema } from "./template.schema";
+import { TemplateDoc, TemplateDocVersion, TemplateSchema } from "./template.schema";
 
 describe("templateRepository", () => {
   let templateRepository: TemplateRepository;
+  let TemplateDocument: MongooseModel<TemplateDoc>;
 
   let module: TestingModule;
 
@@ -42,6 +48,7 @@ describe("templateRepository", () => {
     }).compile();
 
     templateRepository = module.get<TemplateRepository>(TemplateRepository);
+    TemplateDocument = module.get<Model<TemplateDoc>>(getModelToken(TemplateDoc.name));
   });
 
   it("should save a template", async () => {
@@ -53,10 +60,95 @@ describe("templateRepository", () => {
         submodels: [randomUUID()],
         conceptDescriptions: [randomUUID()],
       }),
+      lastStatusChange: DigitalProductDocumentStatusChange.create({
+        previousStatus: DigitalProductDocumentStatus.Draft,
+        currentStatus: DigitalProductDocumentStatus.Published,
+      }),
     });
     await templateRepository.save(template);
     const foundTemplate = await templateRepository.findOneOrFail(template.id);
     expect(foundTemplate).toEqual(template);
+  });
+
+  it(`should load and migrate passport from version 1.0.0 to 1.1.0`, async () => {
+    const id = randomUUID();
+    const now = new Date();
+    const legacyDoc = new TemplateDocument({
+      _id: id,
+      _schemaVersion: TemplateDocVersion.v1_0_0,
+      organizationId: randomUUID(),
+      environment: {
+        submodels: [randomUUID()],
+        assetAdministrationShells: [randomUUID()],
+        conceptDescriptions: [],
+      },
+      createdAt: now,
+      updatedAt: now,
+    });
+    await legacyDoc.save({ validateBeforeSave: false });
+    const foundTemplate = await templateRepository.findOneOrFail(id);
+    expect(foundTemplate).toEqual(
+      Template.fromPlain({
+        id,
+        environment: Environment.fromPlain(legacyDoc.environment),
+        organizationId: legacyDoc.organizationId,
+        createdAt: legacyDoc.createdAt,
+        updatedAt: legacyDoc.updatedAt,
+        lastStatusChange: DigitalProductDocumentStatusChange.create({}),
+      }),
+    );
+  });
+
+  it("should delete a template", async () => {
+    const template = Template.create({
+      id: randomUUID(),
+      organizationId: randomUUID(),
+      environment: Environment.create({
+        assetAdministrationShells: [randomUUID()],
+        submodels: [randomUUID()],
+        conceptDescriptions: [randomUUID()],
+      }),
+    });
+    await templateRepository.save(template);
+    await templateRepository.deleteById(template.id);
+    const foundTemplate = await templateRepository.findOne(template.id);
+    expect(foundTemplate).toBeUndefined();
+  });
+
+  it("should filter templates of version 1.0.0 as draft", async () => {
+    const date1 = new Date("2022-01-01T00:00:00.000Z");
+    const date2 = new Date("2022-02-01T00:00:00.000Z");
+    const organizationId = randomUUID();
+    const createLegacyDoc = async (date: Date) => {
+      const legacyDoc = new TemplateDocument({
+        _id: randomUUID(),
+        _schemaVersion: TemplateDocVersion.v1_0_0,
+        organizationId,
+        environment: {
+          submodels: [],
+          assetAdministrationShells: [],
+          conceptDescriptions: [],
+        },
+        createdAt: date,
+        updatedAt: date,
+      });
+      return await legacyDoc.save({ validateBeforeSave: false });
+    };
+
+    const legacyDoc1 = await createLegacyDoc(date1);
+    const legacyDoc2 = await createLegacyDoc(date2);
+    let foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      filter: {
+        status: [DigitalProductDocumentStatus.Draft],
+      },
+    });
+    expect(foundTemplates.items.map((p) => p.id)).toEqual([legacyDoc2.id, legacyDoc1.id]);
+    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      filter: {
+        status: [DigitalProductDocumentStatus.Published],
+      },
+    });
+    expect(foundTemplates.items.map((p) => p.id)).toEqual([]);
   });
 
   it("should find all templates of organization", async () => {
@@ -67,6 +159,7 @@ describe("templateRepository", () => {
     const date2 = new Date("2022-02-01T00:00:00.000Z");
     const date3 = new Date("2022-03-01T00:00:00.000Z");
     const date4 = new Date("2022-03-02T00:00:00.000Z");
+    const date5 = new Date("2022-03-03T00:00:00.000Z");
 
     const t1 = Template.create({
       id: randomUUID(),
@@ -102,11 +195,29 @@ describe("templateRepository", () => {
         assetAdministrationShells: [randomUUID()],
       }),
       createdAt: date4,
+      lastStatusChange: DigitalProductDocumentStatusChange.create({
+        previousStatus: DigitalProductDocumentStatus.Draft,
+        currentStatus: DigitalProductDocumentStatus.Published,
+      }),
     });
+    const t5 = Template.create({
+      id: randomUUID(),
+      organizationId,
+      environment: Environment.create({
+        assetAdministrationShells: [randomUUID()],
+      }),
+      createdAt: date5,
+      lastStatusChange: DigitalProductDocumentStatusChange.create({
+        previousStatus: DigitalProductDocumentStatus.Draft,
+        currentStatus: DigitalProductDocumentStatus.Archived,
+      }),
+    });
+
     await templateRepository.save(t1);
     await templateRepository.save(t2OtherOrganization);
     await templateRepository.save(t3);
     await templateRepository.save(t4);
+    await templateRepository.save(t5);
 
     let foundTemplates = await templateRepository.findAllByOrganizationId(organizationId);
 
@@ -116,13 +227,48 @@ describe("templateRepository", () => {
           cursor: encodeCursor(t1.createdAt.toISOString(), t1.id),
           limit: 100,
         }),
-        items: [t4, t3, t1],
+        items: [t5, t4, t3, t1],
       }),
     );
+
+    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      filter: {
+        status: [DigitalProductDocumentStatus.Archived],
+      },
+    });
+
+    expect(foundTemplates).toEqual(
+      PagingResult.create({
+        pagination: Pagination.create({
+          cursor: encodeCursor(t5.createdAt.toISOString(), t5.id),
+          limit: 100,
+        }),
+        items: [t5],
+      }),
+    );
+
+    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      filter: {
+        status: [DigitalProductDocumentStatus.Archived, DigitalProductDocumentStatus.Published],
+      },
+    });
+
+    expect(foundTemplates).toEqual(
+      PagingResult.create({
+        pagination: Pagination.create({
+          cursor: encodeCursor(t4.createdAt.toISOString(), t4.id),
+          limit: 100,
+        }),
+        items: [t5, t4],
+      }),
+    );
+
     let pagination = Pagination.create({
       cursor: encodeCursor(t4.createdAt.toISOString(), t4.id),
     });
-    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, pagination);
+    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      pagination,
+    });
     expect(foundTemplates).toEqual(
       PagingResult.create({
         pagination: Pagination.create({ cursor: encodeCursor(t1.createdAt.toISOString(), t1.id) }),
@@ -133,7 +279,9 @@ describe("templateRepository", () => {
       cursor: encodeCursor(t4.createdAt.toISOString(), t4.id),
       limit: 1,
     });
-    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, pagination);
+    foundTemplates = await templateRepository.findAllByOrganizationId(organizationId, {
+      pagination,
+    });
     expect(foundTemplates).toEqual(
       PagingResult.create({
         pagination: Pagination.create({