+++ apps/main/src/unique-product-identifier/unique.product.identifier.module.ts
@@ -17,11 +17,11 @@ import { OrganizationsModule } from "../identity/organizations/organizations.mod
 import { UsersModule } from "../identity/users/users.module";
 import { PassportsModule } from "../passports/passports.module";
 import { TraceabilityEventsModule } from "../traceability-events/traceability-events.module";
+import { UniqueProductIdentifierRepository } from "./infrastructure/unique-product-identifier.repository";
 import {
   UniqueProductIdentifierDoc,
   UniqueProductIdentifierSchema,
 } from "./infrastructure/unique-product-identifier.schema";
-import { UniqueProductIdentifierService } from "./infrastructure/unique-product-identifier.service";
 import { UniqueProductIdentifierApplicationService } from "./presentation/unique.product.identifier.application.service";
 import { UniqueProductIdentifierController } from "./presentation/unique.product.identifier.controller";
 
@@ -48,7 +48,7 @@ import { UniqueProductIdentifierController } from "./presentation/unique.product
     BrandingModule,
   ],
   controllers: [UniqueProductIdentifierController],
-  providers: [UniqueProductIdentifierApplicationService, UniqueProductIdentifierService],
-  exports: [UniqueProductIdentifierService, UniqueProductIdentifierApplicationService],
+  providers: [UniqueProductIdentifierApplicationService, UniqueProductIdentifierRepository],
+  exports: [UniqueProductIdentifierRepository, UniqueProductIdentifierApplicationService],
 })
 export class UniqueProductIdentifierModule {}